fix: update snapshots using pandas to_string()
